backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
google-generativeai==0.3.1
python-docx==1.1.0
python-pptx==0.6.23
python-dateutil==2.8.2
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.1
aiosqlite==0.19.0
greenlet==3.0.1
